@@ -1,16 +1,18 @@
 {
   "name": "noisefy",
-  "version": "0.0.14",
-  "description": "JS library to interact with web audio interface",
+  "version": "0.1.0",
+  "description": "JS library to wrap web audio funny effects",
+  "author": "Arturo Galan <arturo.galan@gmail.com>",
+  "license": "MIT",
   "main": "src/index.js",
   "scripts": {
-    "build": "webpack --env dev && webpack --env build && npm run test",
+    "build": "webpack --env dev && webpack --env build",
     "dev": "webpack --progress --colors --watch --env dev",
     "test": "mocha --require babel-register --colors ./test/*.spec.js",
     "test:watch": "mocha --require babel-register --colors -w ./test/*.spec.js",
     "test:cover": "cross-env NODE_ENV=test nyc mocha --require babel-register --colors test/*.js",
     "repl": "node -i -e \"$(< ./lib/webpack-library-starter.js)\"",
-    "lintfix": "eslint --fix src/**"
+    "lintfix": "eslint --fix src/**/*.js"
   },
   "repository": {
     "type": "git",
@@ -23,40 +25,41 @@
     "guitar",
     "pedal"
   ],
-  "author": "Arturo Galan",
-  "license": "GPL-3.0-or-later",
   "bugs": {
     "url": "https://github.com/arturogalan/noisefy/issues"
   },
   "homepage": "https://github.com/arturogalan/noisefy#readme",
   "devDependencies": {
-    "@babel/cli": "^7.0.0-beta.51",
-    "@babel/core": "^7.0.0-beta.51",
-    "@babel/preset-env": "^7.0.0-beta.51",
+    "@babel/cli": "^7.6.2",
+    "@babel/core": "^7.6.2",
+    "@babel/preset-env": "^7.6.2",
     "babel-eslint": "^8.0.3",
-    "babel-loader": "^8.0.0-beta.4",
+    "babel-loader": "^8.0.6",
     "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-istanbul": "^5.1.0",
+    "babel-plugin-istanbul": "^5.2.0",
     "babel-preset-env": "^7.0.0-beta.3",
     "babel-register": "^7.0.0-beta.3",
     "chai": "^4.1.2",
-    "cross-env": "^5.2.0",
+    "copy-webpack-plugin": "^4.0.1",
+    "cross-env": "^5.2.1",
     "eslint": "^5.0.1",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-import": "^2.16.0",
     "eslint-config-standard": "^11.0.0",
     "eslint-friendly-formatter": "^4.0.0",
+    "eslint-loader": "^2.2.1",
+    "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-node": "^6.0.1",
     "eslint-plugin-promise": "^3.7.0",
     "eslint-plugin-standard": "^3.0.1",
+    "file-loader": "^4.2.0",
     "jsdom": "11.11.0",
     "jsdom-global": "3.0.2",
     "mocha": "^4.0.1",
+    "nyc": "^13.1.0",
     "uglifyjs-webpack-plugin": "^1.2.7",
-    "webpack": "^4.12.2",
-    "webpack-cli": "^3.0.8",
-    "yargs": "^10.0.3",
-    "nyc": "^13.1.0"
+    "url-loader": "^2.1.0",
+    "webpack": "^4.41.0",
+    "webpack-cli": "^3.3.9",
+    "yargs": "^10.0.3"
   },
   "nyc": {
     "sourceMap": false,
