@@ -2,7 +2,7 @@ import SingleAudioNode from '../SingleAudioNode';
 
 /**
  * The audio-effects Basic Presence class.
- * This class lets you add an Presence with bass mid and trebble filters.
+ * This class lets you add a Presence filter.
  */
 export default class Presence extends SingleAudioNode {
   constructor(audioContext) {
