@@ -1,16 +1,37 @@
+
+const DISTORTION_PRESETS = [
+  {name: 'Hard Rock classic 1', distortionStage1: 'asymetric', distortionStage2: 'notSoDistorted'},
+  {name: 'Hard Rock classic 2', distortionStage1: 'standard', distortionStage2: 'notSoDistorted'},
+  {name: 'Hard Rock classic 3', distortionStage1: 'standardLow', distortionStage2: 'notSoDistorted'},
+  {name: 'High Gain 1', distortionStage1: 'asymetric', distortionStage2: 'highGain'},
+  {name: 'High Gain 2', distortionStage1: 'asymetric', distortionStage2: 'highGainModern'},
+  {name: 'Crunch', distortionStage1: 'asymetric', distortionStage2: 'crunch'},
+  {name: 'Fuzz 1', distortionStage1: 'asymetric', distortionStage2: 'fuzz'},
+  {name: 'Fuzz 2', distortionStage1: 'asymetric', distortionStage2: 'superFuzz'},
+];
+
+const CLEAN_PRESETS = [
+  {name: 'Clean and Warm', distortionStage1: 'asymetric', distortionStage2: 'standard'},
+  {name: 'Clean no reverb', distortionStage1: 'asymetric', distortionStage2: 'crunch'},
+  {name: 'Another Clean Sound', distortionStage1: 'asymetric', distortionStage2: 'crunch'},
+  {name: 'Mostly even harmonics', distortionStage1: 'standard', distortionStage2: 'standard'},
+  {name: 'Strong and Warm', distortionStage1: 'asymetric', distortionStage2: 'superClean'},
+  {name: 'SuperClean/Jazz', distortionStage1: 'crunch', distortionStage2: 'crunch'},
+];
+
 const DISTORTION_TYPES = {
-  SUPER_CLEAN: 'superClean',
-  CLEAN: 'clean',
-  SMOOTH: 'smooth',
-  FUZZ: 'fuzz',
-  SUPERFUZZ: 'superFuzz',
-  CRUNCH: 'crunch',
-  NOT_SO_DISTORTED: 'notSoDistorted',
+  ASYMETRIC: 'asymetric',
   STANDARD: 'standard',
   STANDARD_LOW: 'standardLow',
   HIGH_GAIN: 'highGain',
   HIGH_GAIN_MODERN: 'highGainModern',
-  ASYMETRIC: 'asymetric',
+  CLEAN: 'clean',
+  SUPER_CLEAN: 'superClean',
+  SMOOTH: 'smooth',
+  NOT_SO_DISTORTED: 'notSoDistorted',
+  FUZZ: 'fuzz',
+  SUPERFUZZ: 'superFuzz',
+  CRUNCH: 'crunch',
   BEZIER: 'bezier',
   CLASS_A: 'classA',
   VERTICAL: 'vertical',
@@ -29,8 +50,18 @@ const getDistortionTypeGenerateFunction = (type)=> {
       }
       return curve;
     },
+    [DISTORTION_TYPES.STANDARD_LOW]: (intensity)=> {
+      let amount = 44100;
+      let curve = new Float32Array(amount);
+      let deg = Math.PI / 180;
+      for (let i = 0; i < amount; ++i) {
+        let x = i * 2 / amount - 1;
+        curve[i] = (3 + intensity) * x * 20 * deg / (Math.PI + intensity * Math.abs(x));
+      }
+      return curve;
+    },
     [DISTORTION_TYPES.HIGH_GAIN]: (intens)=> {
-      const intensity = intens / 153 * 10;
+      const intensity = (intens / 10) / 153;
       const amount = 22050;
       const curve = new Float32Array(amount);
 
@@ -40,8 +71,7 @@ const getDistortionTypeGenerateFunction = (type)=> {
       return curve;
     },
     [DISTORTION_TYPES.HIGH_GAIN_MODERN]: (intens)=> {
-      let intensity = intens / 2;
-      intensity = 1 / (1 + Math.pow(intensity, 4));
+      let intensity = 1 / (1 + Math.pow((intens / 2), 4));
       const amount = 22050;
       const curve = new Float32Array(amount);
 
@@ -51,13 +81,12 @@ const getDistortionTypeGenerateFunction = (type)=> {
       }
       return curve;
     },
-    // amount not relevant, asymetric fixed curve, usually helpful to generate ODD harmonics
-    [DISTORTION_TYPES.ASYMETRIC]: (intens)=> {
-      let intensity = intens / 1500;
+    // amount not relevant, asymetric fixed curve, helpful to generate ODD harmonics
+    [DISTORTION_TYPES.ASYMETRIC]: ()=> {
       let curve = new Float32Array(44100);
-
-      for (let i = 0; i < intensity; i++) {
-        let x = i * 2 / intensity - 1;
+      const amount = 44100;
+      for (let i = 0; i < amount; i++) {
+        let x = i * 2 / amount - 1;
         if (x < -0.08905) {
           curve[i] = (-3 / 4) * (1 - (Math.pow((1 - (Math.abs(x) - 0.032857)), 12)) + (1 / 3) * (Math.abs(x) - 0.032847)) + 0.01;
         } else if (x >= -0.08905 && x < 0.320018) {
@@ -68,10 +97,160 @@ const getDistortionTypeGenerateFunction = (type)=> {
       }
       return curve;
     },
-  }[type];
+    [DISTORTION_TYPES.NOT_SO_DISTORTED]: (intens)=> {
+      let intensity = Math.pow((intens / 150) + 2, 3);
+      const amount = 22050;
+      const curve = new Float32Array(amount);
+
+      for (let d = 0; amount > d; d += 1) {
+        let f = 2 * d / amount - 1;
+        curve[d] = (1 + intensity) * f / (1 + intensity * Math.abs(f));
+      }
+      return curve;
+    },
+    [DISTORTION_TYPES.SUPER_CLEAN]: (intens)=> {
+      let intensity = (((intens / 150) + 6) / 4);
+      const amount = 22050;
+      const curve = new Float32Array(amount);
+      for (let d = 0; amount > d; d += 1) {
+        let e = 2 * d / 22050 - 1;
+        curve[d] = (1 + intensity) * e / (1 + intensity * Math.abs(e));
+      }
+      return curve;
+    },
+    [DISTORTION_TYPES.CLEAN]: (intens)=> {
+      let intensity = intens / 1500;
+      const amount = 44100;
+      const curve = new Float32Array(amount);
+      let a = 1 - intensity > 0.99 ? 0.99 : 1 - intensity;
+
+      for (let i = 0; i < amount; i++) {
+        let x = i * 2 / amount - 1;
+        let abx = Math.abs(x);
+        let y;
+        if (abx < a) {
+          y = abx;
+        } else if (abx > a) {
+          y = a + (abx - a) / (1 + Math.pow((abx - a) / (1 - a), 2));
+        } else if (abx > 1) {
+          y = abx;
+          curve[i] = sign(x) * y * (1 / ((a + 1) / 2));
+        }
+      }
+      return curve;
+    },
+    [DISTORTION_TYPES.SMOOTH]: (intens)=> {
+      const amount = 44100;
+      let curve = new Float32Array(amount);
+      let intensity = intens / 1500;
+
+      intensity = Math.min(intensity, 0.9);
+      let k = 2 * intensity / (1 - intensity);
+      for (let i = 0; i < amount; i++) {
+        let x = i * 2 / amount - 1;
+        curve[i] = (1 + k) * x / (1 + k * Math.abs(x));
+      }
+      return curve;
+    },
+    [DISTORTION_TYPES.FUZZ]: (intens)=> {
+      let amount = 44100;
+      let curve = new Float32Array(amount);
+      let intensity = intens / 1500;
+      let a = 1 - intensity;
+      for (let i = 0; i < amount; i++) {
+        let x = i * 2 / amount - 1;
+        let y = x < 0 ? - Math.pow(Math.abs(x), a + 0.04) : Math.pow(x, a);
+        curve[i] = tanh(y * 2);
+      }
+      return curve;
+    },
+    [DISTORTION_TYPES.SUPERFUZZ]: (intens)=> {
+      let amount = 22050;
+      let curve = new Float32Array(amount);
+      let intensity = Math.pow((intens / 150), 3);
+      for (let d = 0; amount > d; d += 1) {
+        let e = 2 * d / amount - 1;
+        curve[d] = (1 + intensity) * e / (1 + intensity * Math.abs(e));
+      }
+      return curve;
+    },
+    [DISTORTION_TYPES.CRUNCH]: (intens)=> {
+      let amount = 22050;
+      let curve = new Float32Array(amount);
+      let intensity = Math.pow((intens / 150), 2);
+      for (let d = 0; 22050 > d; d += 1) {
+        let f = 2 * d / 22050 - 1;
+        curve[d] = (1 + intensity) * f / (1 + intensity * Math.abs(f));
+      }
+      return curve;
+    },
+    [DISTORTION_TYPES.BEZIER]: (intens)=> {
+      let p0 = {x: 0, y: 100};
+      let p1 = {x: 10, y: 50};
+      let p2 = {x: 0, y: 50};
+      let p3 = {x: 100, y: 0};
+      let amount = 44100;
+      let accuracy = 1/amount;
+      let curve = new Float32Array(amount);
+      let index = 0;
+      curve[index++] = map(p0.y, 0, 100, 1, -1);
+      for (let i=0; i<1; i+=accuracy) {
+        let p = bezier(i, p0, p1, p2, p3);
+        curve[index++] = map(p.y, 0, 100, 1, -1);
+      }
+      return curve;
+    },
+    [DISTORTION_TYPES.CLASS_A]: (intens)=> {
+      const intensity = 10 + 3 * (intens / 150);
+      const amount = 22050;
+      let curve = new Float32Array(amount);
+      for (let d = 0; amount > d; d += 1) {
+        let e = 2 * d / amount - 1;
+        curve[d] = (1 + intensity) * e / (1 + intensity * Math.abs(e));
+      }
+      return curve;
+    },
+    [DISTORTION_TYPES.VERTICAL]: (intens)=> {
+      const intensity = Math.pow((intens / 150) + 2, 3);
+      const amount = 22050;
+      let curve = new Float32Array(amount);
+      for (let d = 0; 22050 > d; d += 1) {
+        let e = 2 * d / 22050 - 1;
+        curve[d] = (1 + intensity) * e / (1 + intensity * Math.abs(e));
+      }
+      return curve;
+    },
+  }[type] || (()=> {});
+};
+
+
+const sign = (x)=> {
+  if (x === 0) {
+    return 1;
+  }
+  return Math.abs(x) / x;
+};
+const tanh = (n)=> {
+  return (Math.exp(n) - Math.exp(-n)) / (Math.exp(n) + Math.exp(-n));
+};
+const map = (value, istart, istop, ostart, ostop)=> {
+  return ostart + (ostop - ostart) * ((value - istart) / (istop - istart));
+};
+const bezier = (t, p0, p1, p2, p3)=> {
+  let cX = 3 * (p1.x - p0.x);
+  let bX = 3 * (p2.x - p1.x) - cX;
+  let aX = p3.x - p0.x - cX - bX;
+  let cY = 3 * (p1.y - p0.y);
+  let bY = 3 * (p2.y - p1.y) - cY;
+  let aY = p3.y - p0.y - cY - bY;
+  let x = (aX * Math.pow(t, 3)) + (bX * Math.pow(t, 2)) + (cX * t) + p0.x;
+  let y = (aY * Math.pow(t, 3)) + (bY * Math.pow(t, 2)) + (cY * t) + p0.y;
+  return {x: x, y: y};
 };
 
 export {
   DISTORTION_TYPES,
+  DISTORTION_PRESETS,
+  CLEAN_PRESETS,
   getDistortionTypeGenerateFunction,
 };
