@@ -1,7 +1,7 @@
 
+import Cabinet from './audio-nodes/effects/Cabinet';
 import Equalizer from './audio-nodes/effects/Equalizer';
 import Delay from './audio-nodes/effects/Delay';
-import Distortion from './audio-nodes/effects/Distortion';
 import Flanger from './audio-nodes/effects/Flanger';
 import Input from './audio-nodes/effects/Input';
 import Output from './audio-nodes/effects/Output';
@@ -9,10 +9,21 @@ import Presence from './audio-nodes/effects/Presence';
 import Reverb from './audio-nodes/effects/Reverb';
 import Tremolo from './audio-nodes/effects/Tremolo';
 import Volume from './audio-nodes/effects/Volume';
-import {hasAudioContext, hasGetUserMedia, deviceList, deviceListHandler} from './Util';
-import {DISTORTION_TYPES} from './audio-nodes/factories/DistortionGenerator';
+import HighselfFilter from './audio-nodes/effects/HighselfFilter';
+import LowshelfFilter from './audio-nodes/effects/LowshelfFilter';
+import PeakingFilter from './audio-nodes/effects/PeakingFilter';
+import HighpassFilter from './audio-nodes/effects/HighpassFilter';
+
+import {hasAudioContext, hasGetUserMedia, deviceList, deviceListHandler} from './util';
+import Distortion from './audio-nodes/effects/Distortion';
+import {DISTORTION_TYPES, DISTORTION_PRESETS, CLEAN_PRESETS} from './audio-nodes/factories/DistortionGenerator';
+import Amp from './audio-nodes/amp/Amp';
+import {AMP_TYPES, AMP_COMPONENT_NAME, AMP_SETTING_NAME} from './audio-nodes/factories/AmpGenerator';
+import {BIQUAD_FILTER_TYPES} from './audio-nodes/factories/FiltersGenerator';
+import {PRESET_TYPES} from './audio-nodes/factories/PresetGenerator';
 
 export {
+  Cabinet,
   Equalizer,
   Delay,
   Distortion,
@@ -23,9 +34,21 @@ export {
   Reverb,
   Tremolo,
   Volume,
+  HighselfFilter,
+  LowshelfFilter,
+  PeakingFilter,
+  HighpassFilter,
   hasAudioContext,
   hasGetUserMedia,
   deviceList,
   deviceListHandler,
   DISTORTION_TYPES,
+  DISTORTION_PRESETS,
+  CLEAN_PRESETS,
+  Amp,
+  AMP_TYPES,
+  AMP_COMPONENT_NAME,
+  AMP_SETTING_NAME,
+  BIQUAD_FILTER_TYPES,
+  PRESET_TYPES,
 };
